lexical_util::iterator

Struct Bytes

pub struct Bytes<'a, const FORMAT: u128> { /* private fields */ }

Slice iterator that skips characters matching a given value.

This wraps an iterator over a contiguous block of memory, and only returns values that are not equal to skip.

The format allows us to dictate the actual behavior of the iterator: in what contexts does it skip digit separators.

FORMAT is required to tell us what the digit separator is, and where the digit separators are allowed, as well tell us the radix. The radix is required to allow us to differentiate digit from non-digit characters (see DigitSeparators for a detailed explanation on why).

Implementations

impl<'a, const FORMAT: u128> Bytes<'a, FORMAT>

pub const fn new(slc: &'a [u8]) -> Self

Create new byte object.

pub fn integer_iter<'b>(&'b mut self) -> IntegerDigitsIterator<'a, 'b, FORMAT>

Get iterator over integer digits.

pub fn fraction_iter<'b>(&'b mut self) -> FractionDigitsIterator<'a, 'b, FORMAT>

Get iterator over fraction digits.

pub fn exponent_iter<'b>(&'b mut self) -> ExponentDigitsIterator<'a, 'b, FORMAT>

Get iterator over exponent digits.

pub fn special_iter<'b>(&'b mut self) -> SpecialDigitsIterator<'a, 'b, FORMAT>

Get iterator over special floating point values.