lexical_util

Module num

Utilities for Rust numbers.

These traits define useful properties, methods, associated types, and trait bounds, and conversions for working with numbers in generic code.

Traits

AsCast

An interface for casting between machine scalars.

AsPrimitive

Type that can be converted to primitive with as.

Float

Float information for native float types.

Integer

Defines a trait that supports integral operations.

Number

Numerical type trait.

Primitive

Primitive type trait (which all have static lifetimes).

SignedInteger

Defines a trait that supports signed integral operations.

UnsignedInteger

Defines a trait that supports unsigned integral operations.

Functions

as_cast

Allows the high-level conversion of generic types as if as was used.