lexical_util

Module options

Shared traits for the options API.

The following constants have the following signifiers:

${X}_LITERAL: Applies to all literal values for that language (for example, RUST_LITERAL).

${X}_STRING: Applies to all string values for that language (for example, ERLANG_STRING).

${X}: Applies to all values for that language (for example, KAWA).

${X}_(NAN|INF|INFINITY): Applies to only a single special value (for example, PHP_LITERAL_NAN, PHP_LITERAL_INF, and PHP_LITERAL_INFINITY).

If it’s not defined, all values are the default. The default options are:

NaN: (*_NAN): NaN

Short infinity: (*_INF): Inf (including +Inf and -Inf)

Long infinity: (*_INFINITY): Infinity (including +Infinity and -Infinity)

Constants

CLOJURE_LITERAL

A Clojure (Lisp) literal number (uses default options).

CLOJURE_STRING_INF

A Clojure string short infinity (Infinity).

CLOJURE_STRING_INFINITY

A Clojure string long infinity (Infinity).

COBOL

A COBOL literal number (uses default options).

COFFEESCRIPT_INF

A CoffeeScript string short infinity (Infinity).

COFFEESCRIPT_INFINITY

A CoffeeScript string long infinity (Infinity).

CSHARP_LITERAL

A C# literal number (uses default options).

CSHARP_STRING_INF

A C# string short infinity (Infinity).

CSHARP_STRING_INFINITY

A C# string long infinity (Infinity).

CXX_LITERAL_INF

A C++ literal short infinity (INFINITY).

CXX_LITERAL_INFINITY

A C++ literal long infinity (INFINITY).

CXX_LITERAL_NAN

A C++ literal NaN (NAN).

C_LITERAL_INF

A C literal short infinity (INFINITY).

C_LITERAL_INFINITY

A C literal long infinity (INFINITY).

C_LITERAL_NAN

A C literal NaN (NAN).

D_LITERAL

A D-Lang literal number (uses default options).

ELIXIR

An Elixir number (uses default options).

ELM_LITERAL

An Elm literal number (uses default options).

ELM_STRING_INF

An Elm string short infinity (Infinity).

ELM_STRING_INFINITY

An Elm string long infinity (Infinity).

ELM_STRING_NAN

An Elm strong NaN (uses default options).

ERLANG_LITERAL_NAN

An Erlang literal NaN (nan).

ERLANG_STRING

An Erlang string number (uses default options).

FORTRAN_LITERAL

A FORTRAN literal number (uses default options).

FSHARP_LITERAL_INF

An F# literal short infinity (infinity).

FSHARP_LITERAL_INFINITY

An F# literal long infinity (infinity).

FSHARP_LITERAL_NAN

An F# literal NaN (nan).

GAMBITC

A Gambit-C (Lisp) literal number (uses default options).

GO_LITERAL

A Golang literal number (uses default options).

GUILE

A Guile (Lisp) literal number (uses default options).

HASKELL_LITERAL

A Haskell literal number (uses default options).

HASKELL_STRING_INF

A Haskell string short infinity (Infinity).

HASKELL_STRING_INFINITY

A Haskell string long infinity (Infinity).

JAVASCRIPT_INF

A JavaScript string short infinity (Infinity).

JAVASCRIPT_INFINITY

A JavaScript string long infinity (Infinity).

JAVA_LITERAL

A Java literal number (uses default options).

JAVA_STRING_INF

A Java string short infinity (Infinity).

JAVA_STRING_INFINITY

A Java string long infinity (Infinity).

JSON

A JSON number (uses default options).

JULIA_LITERAL_INF

A Julia string short infinity (Inf).

JULIA_LITERAL_INFINITY

A Julia string long infinity (Inf).

KAWA

A Kawa (Lisp) literal number (uses default options).

KOTLIN_LITERAL

A Kotlin literal number (uses default options).

KOTLIN_STRING_INF

A Kotlin string short infinity (Infinity).

KOTLIN_STRING_INFINITY

A Kotlin string long infinity (Infinity).

MATLAB_LITERAL_INF

A MATLAB literal short infinity (inf).

MATLAB_LITERAL_INFINITY

A MATLAB literal long infinity (Inf).

MONGODB_INF

A MongoDB short infinity (Infinity).

MONGODB_INFINITY

A MongoDB long infinity (Infinity).

MYSQL

A MySQL number (uses default options).

OBJECTIVEC

An Objective-C number (uses default options).

OCAML_LITERAL_INF

An OCAML literal short infinity (infinity).

OCAML_LITERAL_INFINITY

An OCAML literal long infinity (infinity).

OCAML_LITERAL_NAN

An OCAML literal NaN (nan).

PERL_LITERAL

A Perl literal literal (uses default options).

PHP_LITERAL_INF

A PHP literal short infinity (INF).

PHP_LITERAL_INFINITY

A PHP literal long infinity (INF).

PHP_LITERAL_NAN

A PHP literal NaN (NAN).

POSTGRESQL

A PostgreSQL number (uses default options).

PYTHON_LITERAL

A Python literal number (uses default options).

REASONML_LITERAL_INF

A ReasonML literal short infinity (infinity).

REASONML_LITERAL_INFINITY

A ReasonML literal long infinity (infinity).

REASONML_LITERAL_NAN

A ReasonML literal NaN (nan).

RUBY_LITERAL_INF

A Ruby literal short infinity (Infinity).

RUBY_LITERAL_NAN

A Ruby literal NaN (NaN).

RUBY_STRING_NONE

A Ruby string (uses default options).

RUST_LITERAL

A Rust literal number (uses default options).

R_LITERAL_INF

An R literal short infinity (Inf).

R_LITERAL_INFINITY

An R literal long infinity (Inf).

SAGE_LITERAL_INF

A SageMath literal short infinity (infinity).

SAGE_LITERAL_INFINITY

A SageMath literal long infinity (Infinity).

SCALA_LITERAL

A Scala literal number (uses default options).

SCALA_STRING_INF

A Scala string short infinity (Infinity).

SCALA_STRING_INFINITY

A Scala string long infinity (Infinity).

SQLITE

A SQLite number (uses default options).

SWIFT_LITERAL

A Swift literal number (uses default options).

TOML

A TOML number (uses default options).

VB_LITERAL

A Visual Basic literal number (uses default options)

VB_STRING_INF

A Visual Basic short string infinity (uses default options)

VB_STRING_INFINITY

A Visual Basic long string number (uses default options)

XML_INF

An XML short infinity (uses default options).

XML_INFINITY

An XML short infinity (uses default options).

YAML

A YAML number (uses default options).

ZIG_LITERAL

A Zig literal number (uses default options).

Traits

ParseOptions

Shared trait for all parser options.

WriteOptions

Shared trait for all writer options.