lexical_util::step

Function max_step

pub const fn max_step(radix: u32, bits: usize, is_signed: bool) -> usize

Get the maximum number of digits that can be processed without overflowing.

Calculate the number of digits that can be processed without overflowing for a given type, that is, it can process at least radix^N. This does not necessarily mean it can process every value in the range [0, radix^(N+1)).

For example, with u8 with radix 10, we have a maximum value of 255, so we have a max step of 3: that is, it can process up to 3 digits ([100, 256)), even if it cannot process every 3 digit value ([256, 1000)).