Struct lexical_util::iterator::Bytes

pub struct Bytes<'a, const __: u128> { /* private fields */ }

Slice iterator that stores the original length of the slice.

Implementations

impl<'a, const __: u128> Bytes<'a, __>

pub const IS_CONTIGUOUS: bool = true

If each yielded value is adjacent in memory.

pub const fn new(slc: &'a [u8]) -> Self

Create new byte object.

pub fn as_ptr(&self) -> *const u8

Get a ptr to the current start of the iterator.

pub fn as_slice(&self) -> &'a [u8]

Get a slice to the current start of the iterator.

pub fn length(&self) -> usize

Get the total number of elements in the underlying slice.

pub fn cursor(&self) -> usize

Get the current index of the iterator in the slice.

pub unsafe fn set_cursor(&mut self, index: usize)

Set the current index of the iterator in the slice.

Safety

Safe if index <= self.length().

pub fn current_count(&self) -> usize

Get the current number of values returned by the iterator.

pub fn is_done(&self) -> bool

Get if the buffer underlying the iterator is empty. Same as is_consumed.

pub unsafe fn read_unchecked<V>(&self) -> V

Read a value of a difference type from the iterator. This advances the internal state of the iterator.

Safety

Safe as long as the number of the buffer is contains as least as many bytes as the size of V.

pub fn read<V>(&self) -> Option<V>

Try to read a value of a different type from the iterator. This advances the internal state of the iterator.

pub fn first_is(&mut self, value: u8) -> bool

Check if the next element is a given value.

pub fn case_insensitive_first_is(&mut self, value: u8) -> bool

Check if the next element is a given value without case sensitivity.

pub fn integer_iter<'b>(&'b mut self) -> BytesIterator<'a, 'b, __>

Get iterator over integer digits.

pub fn fraction_iter<'b>(&'b mut self) -> BytesIterator<'a, 'b, __>

Get iterator over fraction digits.

pub fn exponent_iter<'b>(&'b mut self) -> BytesIterator<'a, 'b, __>

Get iterator over exponent digits.

pub fn special_iter<'b>(&'b mut self) -> BytesIterator<'a, 'b, __>

Get iterator over special floating point values.

pub unsafe fn step_by_unchecked(&mut self, count: usize)

Advance the byte by N elements.

Safety

As long as the iterator is at least N elements, this is safe.

pub unsafe fn step_unchecked(&mut self)

Advance the byte by 1 element.

Safety

Safe as long as the iterator is not empty.